rename to MessageBus::Redis::ReliablePubSub
